Add a request callback and a method of adding other callbacks
This is useful for things such as allowing the user to count the number of requests being made to Xero (important due to the very low 1000 requests per day limit), log requests being made, or it could be used to test that the users code is resulting in the desired requests. The use is really up to the imagination of the person writing the callback. I have deliberately kept this quite for now.